refactor: move streaming utils into separate file
